pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "wingman"
version = "1.0.0"
description = "Multi-platform monitoring agent for GitHub, SPT-Forge, and Reddit"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.5.0",
    "praw>=7.8.0",
    "anthropic>=0.40.0",
    "APScheduler>=3.10.0,<4.0.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
browser = ["playwright>=1.49.0"]
dev = ["pytest>=8.0.0", "pytest-mock>=3.14.0", "responses>=0.25.0"]

[tool.setuptools.packages.find]
include = ["wingman*"]
